Extract handlers to `topic_handlers` module
