+++ hermes/bin/src/ipfs/task.rs
@@ -1,17 +1,13 @@
 //! IPFS Task
 use std::{
     collections::HashSet,
-    pin::Pin,
     str::FromStr,
     sync::{Arc, Mutex},
     time::Duration,
 };
 
 use catalyst_types::smt::Tree;
-use hermes_ipfs::{
-    Cid, HermesIpfs, IpfsPath as PathIpfsFile, PeerId as TargetPeerId,
-    doc_sync::payload::{self},
-};
+use hermes_ipfs::{Cid, HermesIpfs, IpfsPath as PathIpfsFile, PeerId as TargetPeerId};
 use tokio::{
     sync::{mpsc, oneshot},
     task::JoinHandle,
@@ -21,15 +17,15 @@ use tokio::{
 use super::HERMES_IPFS;
 use crate::{
     app::ApplicationName,
-    ipfs::{doc_sync::handle_doc_sync_topic, topic_message_context::TopicMessageContext},
+    ipfs::{
+        topic_handlers::{self, TopicMessageHandler, TopicSubscriptionStatusHandler},
+        topic_message_context::TopicMessageContext,
+    },
     runtime_extensions::{
         bindings::hermes::ipfs::api::{
-            DhtKey, DhtValue, Errno, IpfsFile, MessageData, PeerId, PubsubMessage, PubsubTopic,
+            DhtKey, DhtValue, Errno, IpfsFile, MessageData, PeerId, PubsubTopic,
         },
-        hermes::{
-            doc_sync::{self, OnNewDocEvent},
-            ipfs::event::OnTopicEvent,
-        },
+        hermes::doc_sync::{self, OnNewDocEvent},
     },
     wasm::module::ModuleId,
 };
@@ -192,14 +188,14 @@ pub(crate) async fn ipfs_command_handler(
                     SubscriptionKind::Default => {
                         TopicMessageHandler::new(
                             &topic,
-                            topic_message_handler,
+                            topic_handlers::topic_message_handler,
                             TopicMessageContext::new(None, app_name, module_ids),
                         )
                     },
                     SubscriptionKind::DocSync => {
                         TopicMessageHandler::new(
                             &topic,
-                            doc_sync_topic_message_handler,
+                            topic_handlers::doc_sync_topic_message_handler,
                             TopicMessageContext::new(tree, app_name, module_ids),
                         )
                     },
@@ -290,144 +286,6 @@ pub(crate) async fn ipfs_command_handler(
         tracing::warn!("Could not stop IPFS node - other references still exist");
     }
     Ok(())
-}
-
-/// A handler for messages from the IPFS pubsub topic
-// TODO[rafal-ch]: Move to a dedicated module
-#[derive(Clone)]
-pub(super) struct TopicMessageHandler {
-    /// The topic.
-    topic: String,
-
-    /// The handler implementation.
-    #[allow(
-        clippy::type_complexity,
-        reason = "to be revisited after the doc sync functionality is fully implemented as this type still evolves"
-    )]
-    callback: Arc<
-        dyn Fn(
-                hermes_ipfs::rust_ipfs::GossipsubMessage,
-                String,
-                TopicMessageContext, /* TODO[rafal-ch]: Should become a borrow, but if not
-                                      * possible, at least an Arc */
-            ) -> Pin<Box<dyn Future<Output = ()> + Send>>
-            + Send
-            + Sync
-            + 'static,
-    >,
-
-    /// The context.
-    context: TopicMessageContext,
-}
-
-impl TopicMessageHandler {
-    /// Creates the new handler.
-    pub fn new<F, Fut>(
-        topic: &str,
-        handler: F,
-        context: TopicMessageContext,
-    ) -> Self
-    where
-        F: Fn(hermes_ipfs::rust_ipfs::GossipsubMessage, String, TopicMessageContext) -> Fut
-            + Send
-            + Sync
-            + 'static,
-        Fut: Future<Output = ()> + Send + 'static,
-    {
-        Self {
-            topic: topic.to_string(),
-            callback: Arc::new(move |msg, topic, ctx| Box::pin(handler(msg, topic, ctx))),
-            context,
-        }
-    }
-
-    /// Forwards the message to the handler.
-    pub async fn handle(
-        &self,
-        msg: hermes_ipfs::rust_ipfs::GossipsubMessage,
-    ) {
-        (self.callback)(msg, self.topic.clone(), self.context.clone()).await;
-    }
-}
-
-/// A handler for subscribe/unsubscribe events from the IPFS pubsub topic
-#[derive(Clone)]
-pub(super) struct TopicSubscriptionStatusHandler<T>
-where T: Fn(hermes_ipfs::SubscriptionStatusEvent, String) + Send + Sync + 'static
-{
-    /// The topic.
-    topic: String,
-
-    /// The handler implementation.
-    callback: T,
-}
-
-impl<T> TopicSubscriptionStatusHandler<T>
-where T: Fn(hermes_ipfs::SubscriptionStatusEvent, String) + Send + Sync + 'static
-{
-    /// Creates the new handler.
-    pub fn new(
-        topic: &impl ToString,
-        callback: T,
-    ) -> Self {
-        Self {
-            topic: topic.to_string(),
-            callback,
-        }
-    }
-
-    /// Passes the subscription event to the handler.
-    pub fn handle(
-        &self,
-        subscription_event: hermes_ipfs::SubscriptionStatusEvent,
-    ) {
-        (self.callback)(subscription_event, self.topic.clone());
-    }
-}
-
-/// Handler function for topic message streams.
-async fn topic_message_handler(
-    message: hermes_ipfs::rust_ipfs::GossipsubMessage,
-    topic: String,
-    context: TopicMessageContext,
-) {
-    if let Some(ipfs) = HERMES_IPFS.get() {
-        let app_names = ipfs.apps.subscribed_apps(SubscriptionKind::Default, &topic);
-
-        drop(
-            OnTopicEvent::new(PubsubMessage {
-                topic,
-                message: message.data.into(),
-                publisher: message.source.map(|p| p.to_string()),
-            })
-            .build_and_send(app_names, context.module_ids()),
-        );
-    } else {
-        tracing::error!("Failed to send on_topic_event. IPFS is uninitialized");
-    }
-}
-
-/// Handler for Doc Sync `PubSub` messages.
-#[allow(
-    clippy::needless_pass_by_value,
-    reason = "the other handler consumes the message and we need to keep the signatures consistent"
-)]
-async fn doc_sync_topic_message_handler(
-    message: hermes_ipfs::rust_ipfs::GossipsubMessage,
-    topic: String,
-    context: TopicMessageContext,
-) {
-    if let Ok(msg_str) = std::str::from_utf8(&message.data) {
-        tracing::info!(
-            "RECEIVED PubSub message on topic: {topic} - data: {}",
-            &msg_str.chars().take(100).collect::<String>()
-        );
-    }
-
-    let result = handle_doc_sync_topic::<payload::New>(&message, &topic, context).await;
-    if let Some(Err(err)) = result {
-        tracing::error!("Failed to handle IPFS message: {}", err);
-    }
 }
 
 /// Processes the received CIDs from a broadcasted message.
